Surface Vanguard command-zone abilities through server view and client
Server: `KnownCard.zone_abilities` carries the activated abilities a card can
be used for from the zone it currently occupies — today, a Vanguard avatar's
`from_command_zone` abilities. Empty for every other card.

Client: a click on your own command-zone card now checks `zone_abilities`
first and submits `ActivateAbility` (arming the targeting cursor when the
ability targets) instead of `CastFromCommandZone`, which a Vanguard can never
legally take.

Also: two `KnownCard` doc comments had drifted onto the wrong fields, and two
clippy lints from the previous commit.

+++ crabomination/src/game/effects/mod.rs
@@ -20223,7 +20223,7 @@ impl GameState {
 
             Effect::LoseAllButLifeRemembered { who, keep } => {
                 let Some(p) = self.resolve_player(who, ctx) else { return Ok(()) };
-                let keep = self.evaluate_value(keep, ctx).max(0) as i32;
+                let keep = self.evaluate_value(keep, ctx).max(0);
                 let lost = self.players[p].life - keep;
                 if lost > 0 {
                     let applied = self.adjust_life_applied(p, -lost);
